Passing Venue ID, Saving Student data to database
Another big day of success,
Passing the proper venue id for the event so it is no longer hard coded
in place. Also when the student tries to check in, we now save their
name, and email up to the database. Also compare against to see if they
already exists, if they do we return the result back instead of
duplicating students