Add Event Page Set Up
Full Cycle for student check-in with email authorization as well as the
Add New Event Page working.

We can now add new events from the page and posts them to the database.

Some DB changes had to be made and CISPROD will be updated with them as
well.

I will export a clean SQL and email and add to OneDrive